+++ src/main/java/com/dwarfeng/ftp/configuration/SingletonConfiguration.java
@@ -17,6 +17,15 @@ import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 @Configuration
 public class SingletonConfiguration {
 
+    public static final String SPEL_DEFAULT_TEMPORARY_FILE_DIRECTORY_PATH = "${ftp.temporary_file_directory_path:" +
+            "#{T(com.dwarfeng.ftp.struct.FtpConfig$Builder).DEFAULT_TEMPORARY_FILE_DIRECTORY_PATH}}";
+    public static final String SPEL_DEFAULT_TEMPORARY_FILE_PREFIX = "${ftp.temporary_file_prefix:" +
+            "#{T(com.dwarfeng.ftp.struct.FtpConfig$Builder).DEFAULT_TEMPORARY_FILE_PREFIX}}";
+    public static final String SPEL_DEFAULT_TEMPORARY_FILE_SUFFIX = "${ftp.temporary_file_suffix:" +
+            "#{T(com.dwarfeng.ftp.struct.FtpConfig$Builder).DEFAULT_TEMPORARY_FILE_SUFFIX}}";
+    public static final String SPEL_DEFAULT_FILE_COPY_MEMORY_BUFFER_SIZE = "${ftp.file_copy_memory_buffer_size:" +
+            "#{T(com.dwarfeng.ftp.struct.FtpConfig$Builder).DEFAULT_FILE_COPY_MEMORY_BUFFER_SIZE}}";
+
     private final ThreadPoolTaskScheduler scheduler;
 
     @Value("${ftp.host}")
@@ -43,6 +52,33 @@ public class SingletonConfiguration {
     @Value("${ftp.buffer_size:#{T(com.dwarfeng.ftp.struct.FtpConfig$Builder).DEFAULT_BUFFER_SIZE}}")
     private int bufferSize;
 
+    /**
+     * @since 1.2.0
+     */
+    // SPEL 太长，故使用常量缩短长度。
+    @Value(SPEL_DEFAULT_TEMPORARY_FILE_DIRECTORY_PATH)
+    private String temporaryFileDirectoryPath;
+
+    /**
+     * @since 1.2.0
+     */
+    // SPEL 太长，故使用常量缩短长度。
+    @Value(SPEL_DEFAULT_TEMPORARY_FILE_PREFIX)
+    private String temporaryFilePrefix;
+    /**
+     * @since 1.2.0
+     */
+    // SPEL 太长，故使用常量缩短长度。
+    @Value(SPEL_DEFAULT_TEMPORARY_FILE_SUFFIX)
+    private String temporaryFileSuffix;
+
+    /**
+     * @since 1.2.0
+     */
+    // SPEL 太长，故使用常量缩短长度。
+    @Value(SPEL_DEFAULT_FILE_COPY_MEMORY_BUFFER_SIZE)
+    private int fileCopyMemoryBufferSize;
+
     @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
     public SingletonConfiguration(ThreadPoolTaskScheduler scheduler) {
         this.scheduler = scheduler;
@@ -51,7 +87,8 @@ public class SingletonConfiguration {
     @Bean(initMethod = "start", destroyMethod = "stop")
     public FtpHandler ftpHandler() {
         FtpConfig ftpConfig = new FtpConfig(
-                host, port, username, password, serverCharset, connectTimeout, noopInterval, bufferSize
+                host, port, username, password, serverCharset, connectTimeout, noopInterval, bufferSize,
+                temporaryFileDirectoryPath, temporaryFilePrefix, temporaryFileSuffix, fileCopyMemoryBufferSize
         );
 
         return new FtpHandlerImpl(scheduler, ftpConfig);