+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - `FtpHandler` 增加新方法。
   - FtpHandler.clearDirectory。
+  - FtpHandler.copyFile。
 
 - 优化部分类的文档注释。
   - com.dwarfeng.ftp.handler.FtpHandler。